pokedex/requirements/yolo_requirements.txt
# YOLO Training Requirements
# For YOLOv3 training with 1025 classes in Google Colab

# Core YOLO framework
ultralytics>=8.0.0

# PyTorch ecosystem
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0

# Experiment tracking
wandb>=0.15.0

# Dataset handling
datasets>=2.0.0
huggingface_hub>=0.15.0

# Computer vision
opencv-python>=4.5.0
pillow>=8.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.60.0

# Note: google-colab package is pre-installed in Colab environment
# Do not install it manually as it can conflict with the system version 